Remove partition accessor functions
I've changed my mind about these. The ones that return pointers to heap
especially. The worry is that the returned pointer will be kept after
the object is destroyed. Then there's a pointer dangling to freed memory.
Better to access the fields directly. Then it is obvious what is being
accessed (the object) and whether it is still around or not.